requirements.txt
# Discord and Networking
discord.py>=2.3.0
aiohttp>=3.8.0
websocket-client>=1.7.0
requests>=2.31.0
urllib3>=2.0.7

# AI and Machine Learning
scikit-learn>=1.0.2
numpy>=1.21.0
openai>=1.12.0
huggingface_hub[hf_transfer]>=0.21.3
google-generativeai>=0.3.0
anthropic>=0.3.0


# Database and Storage
SQLAlchemy>=1.4.23
redis>=4.0.0
pymilvus>=2.3.0

# Image Processing
Pillow>=8.3.2

# Utilities and Tools
python-dotenv>=1.0.0
wget==3.2
pathlib==1.0.1
typing==3.7.4.3
typing-extensions>=4.8.0
tqdm>=4.66.1
watchdog==6.0.0
certifi>=2024.2.2
platform-utils>=0.3.0

# Web Server
gunicorn>=20.1.0

# GUI
ttkthemes>=3.2.2

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.20.0
